Added the WFG MOO Benchmark Set.
Added the Walking Fish Group (WFG) MOO Benchmark Set. These set of
functions were directly ported from the official C++ implementation. The
result of this is that the port is very ugly and needs to be refactored
at some point. Also, to verify the port a set of data files were
generated with their utility. The tests read these data files and
attempts to generate the same results. Error values are checked for all
the x values.