Generalized moving peaks benchmark function.
Revised implementation of the MovingPeaks benchmark problem. The problem
has been generalized to allow for any number of peaks in n dimensions.

DynamicOptimizationProblem has been introduced to allow for the definition
of problem maximums and minimums to allow for the calculation of error
measurements for input vectors to dynamic environments. A sample xml file
demonstrates how a charged PSO can be used to optimize the MovingPeaks
function. Some of the test cases have been migrated to use Mockito
instead of JMock as the syntax is much, much more clear.

Signed-off-by: Bennie Leonard <bennie.leonard@gmail.com>
Signed-off-by: Gary Pampara <gpampara@gmail.com>