Simplify event publishing.
The @Raises caused some concerns as it implies that the provided event
will alwyas be raised. To prevent this confusion, we have changed the
annotation to be @CanRaise to indicate that it is optional.

The remainder is cleanup to the code to make it more readable and
consice as it is getting rather convoluted.

Signed-off-by: Gary Pampara <gpampara@gmail.com>