Remove SableCC -> rather use Parboiled.
The Parboiled parser has the benefit that it is defined in code directly
and does not need an external process to generate the parser.
Additionally, using Parboiled also means that a number of classes are no
longer required in the source.

The maven assembly plugin is also something that is not needed anymore
and should be removed completely.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>