New DE strategy and MovingPeaks bug-fix
A new entity creation strategy is provided for use with DE. The new strategy
enables the creation of trial vectors by selecting new participants from
the population for each dimension when the difference vector is computed.

Additionally, a fix is provided for the GeneralizedMovingPeaks benchmark so
that the arrays used to control peak movements are correctly re-initialized
when the problem dimension is set via xml.

Signed-off-by: Bennie Leonard <bennie.leonard@gmail.com>
Signed-off-by: Gary Pamparà <gpampara@gmail.com>