Entity Creation Strategy Modification
Changed the way in which trial vectors are created. Instead of passing a fixed scale
to the parameter to the Vector class, the Supplier of the ControlParameter is passed.
The effect of this change is that each dimension of the difference vector will be scaled
separately by querying the Supplier for a new value. The complexity of the operation is
not increased, because previously a Supplier of constant values was generated and
queried as opposed to the Supplier of the ControlParameter itself.

The functionality of DE remains unchanged when using a ConstantControlParameter, but
becomes much more flexible when considering a RandomControlParameter, for example.