Entity Creation Strategy Modification
Changed the way in which trial vectors are created. Instead of passing
a fixed scale to the parameter to the Vector class, the Supplier of the
ControlParameter is passed.

The effect of is that each dimension of the difference vector will be scaled
separately by querying the Supplier for a new value. The complexity of the
operation is not increased, as previously a Supplier of constant value
was generated and queried, as opposed to the Supplier of the
ControlParameter itself.

Signed-off-by: Bennie Leonard <bennie.leonard@gmail.com>
Signed-off-by: Gary Pamparà <gpampara@gmail.com>