Fixed possible bug in ForwardingLayer.getActivations()
In the previous implementation, thisSize will always have a value of 1. As a result, only one neuron from the current layer is added to the Vector.
I'm a bit uncertain of why this hasn't caused problems. Both the previous and the new implementations pass unit testing.

The new implementation now makes use of functionality that is already implemented in getNeuralInput().