Added framework for the Cascade architecture.
This includes  a builder for the cascading hidden layer, a builder for the
overall Cascade architecture and a visitor that propagates input through
Cascade architecture. A method to grow the network and incrementally
train it is still required.

A get method was added to NeuralInputSource and its associated
implementation classes to allow for the construction of a consolidated
layer in CascadeVisitor.