Corrected BitStringMatcher and PSO.
I do not understand why the initial setting of fitnesses was wrapped in
a try/catch. I have removed it as it is completely wrong to swallow
an exception. Honestly, using exceptions is not the best part of Java.

The Game playing code needs serious help right now. I'm not happy with it at
all and it is extremely inconsistent. If its broken right now, so be it.
I'd rather fix it properly instead of pretending that it actually works.
In the larger picture, I'm unconcerned as the SBT build needs more priority
and the migration of the code to use the Scala DSL.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>