Allow for domain information to be stored in a NN.
Each Neuron now keeps track of its own domain. Layer calculates its
domain by concatenating the domains of the Neurons in it.
Architecture calculates its domain by concatenating the domains of
the Layers in it.

PrototypeFullyConnectedLayerBuilder now assigns a domain to each
Neuron. It is also now capable of calculating a Neuron's domain
based on its fanin.