Added NegativeOfControlParameter.
This wrapper simply negates the ControlParameter that it wraps.

The bounded getParameter method has not been implemented because
I'm not certain of how the bounds are supposed to behave during
negation. I recommend that it remains as such until someone needs
it. That person is more likely to know how it's supposed to behave.