Added LambdaGammaNeuronDomain.
The new lambda-gamma domain provider sets one lambda and one gamma
domain for each neuron. Since this is done for each neuron instead
of an entire network, it implies that NN domains are a bit
different, but shouldn't affect any algorithms.

LambdaGammaVisitor and LambdaGammaSolutionConversionStrategy has
been adapted to deal with the new domain organisation.

Some unit tests had to be adapted.

Inadvertently fixed a bug: LambdaGammaSolutionConvertionStrategy
needs to be initialised, but the initialise method is not called
anywhere. The new implementation doesn't need initialisation.
Initialise method removed as it doesn't do anything.

Some code cleanup.