Added behaviour to Entity/AbstractEntity.
The behaviour is effectively a per-entity single algorithm iteration.
The idea is that all algorithms will eventually be implemented as
behaviours and an entity's type is determined by the behaviour
assigned to it.

ParticleBehaviour is removed, as it is now obsolete.

Added StandardParticleBehaviour. This is the behaviour most common
in the PSO field.

Modified a few of the PSO iteration strategies to use performIteration
in StandardParticleBehaviour.

Modified much code to use the new StandardParticleBehaviour in much
the same way they used ParticleBehaviour.

Updated the XML files to use the new Behaviour.