Simplify stopping condition usage.
The stopping conditions where implemented in a way that caused a fair
amount of confusion. This has now been simplified and the need for the
google guava predicate class has been removed, preferring the FJ F class
instead. Additionally, the evaluation of the stopping conditions is now
done using the boolean disjunction monoid and the List applicative
functor.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>