Moved fitness evalution from iteration strategies to behaviours.
Fitness calculator is now set in AbstractBehaviour rather than
AbstractEntity.
calculateFitness has been replaced by updateFitness, which forces
the fitness to be calculated externally. Code that calls calculate-
Fitness has been changed accordingly in many places.

Added DoNothingBehaviour. This acts mostly as a stopgap for algorithms
that do not yet have behaviours. Also added this behaviour in the
constructors of entities that need it.