Total function: RVar#choices
If the requested number of elements to select is more than what can be
selected from the provided list, then return a None value. As a result
of this, the function is now total (in my mind) as it will always
terminate and the resulting value is logical by either working or not
working without any "magic".

Signed-off-by: Gary Pamparà <gpampara@gmail.com>