Update GCPSO to apply needed parameter updates.
Even though these changes are needed - there is still the issue that
the velocity needs to be updated in the state of the Entity. Right now
the velocity value of the Entity parameters is being ignored and needs
to be updated to ensure that the algorithm is correct. The question
comes in on how exactly to do this - there are lenses involved here and
I'm sure that there has to be a nicer way to update the value the lens
is pointing to in the parameter structure.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>