Prefer the use of IO.
We need to isolate the creation of side-effects and track them
carefully. This means that we can ask the compiler to help us
with the process.

Furthermore, we are lucky enough to have a REPL where we can play
with and debug portions of logic, so printlns in the code are not
really needed. For explicit tracing, `Writer` can be used.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>