Added implementation for Iteration#async
The entire Iteration level computation needs to be looked at. I can't
really describe why, but the feeling I get from the code is an uneasy
one. I think that the implementations can be much simpler and hopefully
such that the inference is better. Additionally, I keep contemplating
the use of a Free+Coyoneda to get the `Iteration` level implemented.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>