Enable relative fitness based on position context.
The position of an Entity determines the fitness for the current
problem definition. In the case of cooperative algorithms, the
fitness is determined by evaluating a context position, thereby
making the fitness of the current entity position relative to
the built up context position. This is furthermore, a special
case of normal evaluation, where the context is the current
entity position.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>