Simplify usages with non-empty Positions.
The `Position` could take any type constructor, needing the correct
annotations through out the code base in order for the type constructor
to line up. This is overly specified, as a given position in a defined
search space can never be *empty*. Search spaces are defined using
`Interval`s, so we know for a fact that the search spaces that algorithms
can operate on have a dimensionality of at least 1.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>