Replace Reader with plain function
Reader as a data structure doesn't provide too much functionality
for us at the moment. It might be reconsidered in the future, but
for now it is not neccessary.

Split `Position` into its own file and did some minor cleanup based
on the Reader removal. Additionally, added generic `Positive` and
`Negative` types that are constrained to `Numeric`s.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>