Remain in RVar and additional selection types
Some variants for selection where added so that the functions
need not be forced to have a shape conforming to all the other
selections, which is far better for reuse purposes. For the
crossover selections etc, the computation need not be lifted into
`Step` as it can remain within `RVar`, lifted only where needed.

Some minor updates where made to `Guides` to cater for the changes
with crossovers.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>