Add example of GCPSO
There was no exmaple of using an algorithm that uses a `StepS`
structure internally. The GCPSO is, however, such an algorithm
and the added example shows the usage to be as simple as other
algorithms. The only additional concern is the set of algorithm
parameters that the algorithm itself needs to execute.

These additional parameters are passed in first, allowing the
computation to yield a `Step` which then runs the same as in other
examples.

Signed-off-by: Gary Pamparà <gpampara@gmail.com>