Add wart detection
There are a few incorrect usages that need to be prevented to ensure
that the library remains not only type-safe, but logically
sound. Luckily there is a detection library that we can use to ensure
that incorrect and _possibly_ incorrect usage is prevented, as the
scala stdlib is not as safe as what would be expected.