Ensure all entity collections have size >= 1
It was possible to create a collection with _any_ integer value size,
but this is not really something that makes sense. To prevent this
usage error, the parameter to the collection creation functions now
enforce that the input parameter is a "positive integer" (i.e.: the
set of integers >= 1).

If at all possible, the value will be checked at compile time, but in
the cases where the value needs to be determined at runtime, a runtime
check will be done to ensure that the constraint (n >= 1) still holds.

As it is not possible to really recover from this state at runetime,
the program will be terminated, with the appropriate error message.