Do not hardcode entity state parameter in DE and GA
The state parameter for the Entity was hardcoded in the GA and DE to
be `Unit`. This worked to get the implementation done, but there are
problems with keeping it paritally applied - if the user would like to
combine different algorithms into a pipeline, this would prevent such
usage as the Entity state type would be fixed.