Environment product type to clean Step usage
The Step was currently taking two distinct parameters. In situations
where we are experimenting with the removal of the boundary member on
Position, this would result in yet another paramter to the function on
Step. To make this a little cleaner, a product type called
"Environment" has been added that collects the needed values and then
is used as the input to the Step transformer. It is essentially the
environment value for the ReaderT transformer (which Step is).