Stream-based runner (#269)
* Stream-based runner

Additionally, restricted the possible Monad instances that can be
used with the `repeat` function and also generalized `repeat` to work
for both `Step` and `StepS` structures.

* Allow for custom measurements

Measurements that are collected from a collection and then saved
as a case class, will automatically have a parquet schema derived
and as a result will be able to be saved to disk in parquet format.

* Force string identifiers for algorithm and problem

The user is now forced to define a name for the algorithm and the
problem so that they can be identified in the result output. It's not
the best strategy, but it is a valid one until a better scheme is
devised.

* Disallow empty algorithm and problem names

Move empty literals for algorithm and problem names to compile time
errors.

* Adjust the collection after environment changes

Missed acutally using the `onChange` handler in the interpretter.