Repository admin (#272)
* Update repository admin

The LICENSE file has been updated to the Apache-2.0 license. It's
probably the most popular license for OSS software and is essentially
the same as the MIT/BSD licenses.

The CONTRIBUTING file has been updated to allow for referencing the
new _master_ branch, instead of 'series/2.0.x'. Seeing as all
development will be done against this branch anyway, we might as well
keep the name of the branch simple.

Lastly, an initial CHANGELOG file has been added which will detail the
changes of the project from version to version.