Rename point -> pure and pointR -> liftR (#275)
As within Haskell, pure is the preferred usage. It also makes a little
more sense as the provided value is placed into the monad context such
that the same value will be extracted later. Similarly, the same
argument can be made for the pointR -> liftR rename, as you are
lifting a random value (RVar) into another context.