Algorithm stream, allowing for changes based on iteration number
The one feature required to complete the runner was the ability to
have the algorithm parameters vary over the course of the
execution. This patch introduces such a function that allows the user
to provide all the needed logic and produces a stream of algorithm
instances from the prodiced data.

Fixes #217