Adjust reflect to consider the offending portion only
The reflect strategy behaves a bit strangely. It will work
correctly until the value is: > 2*range of domain

If the value is larger, the resulting value will _still_ be
outside the feasible range of the problem space. This change will
essentially ignore and then shrink the value so that it is within
the 2*range before doing the reflect back into the feasible space.
It is not clear if this is _valid_, however, in the normal usage case
this shrinking behaviour would never be done.