Simplified boundary enforcement (#283)
* Added a few bchms

* Generic methods

* Formating

* flatten  for tuples

* Moved to a nchm object in Entity

* Removed tuple flatten

* scalafmt

* No partial functions

* if statements

* Simplify boundary constraints to be plain functions

* Cleanup (more) of functions

* Some more cleanup

* Simplify bondary enforcement to a single function

* Adjust structures to make inference by compiler better

This new scheme now allows for a context for a given boundary handling
mechanism. Some of the strategies require that the new value is
randomly initialized between the domain bounds.

* Correct deployment branch: series/2.0.x -> master

* Add tests for boundary enforcement and correct toroidal scheme

* Adjust reflect to consider the offending portion only

The reflect strategy behaves a bit strangely. It will work
correctly until the value is: > 2*range of domain

If the value is larger, the resulting value will _still_ be
outside the feasible range of the problem space. This change will
essentially ignore and then shrink the value so that it is within
the 2*range before doing the reflect back into the feasible space.
It is not clear if this is _valid_, however, in the normal usage case
this shrinking behaviour would never be done.

* Correct formatting

* Continuous reflecting until within range

* Scalafmt changes

Co-authored-by: Kyle Erwin <kyle.erwin24@gmail.com>