Update Fitness tests and code to check against other cases
The objective code was only considering data that _had_ a fitness
value already assigned. Unfortunately this is a partial solution so
this commit adds the needed changes to allow `Infeasible`, `Adjusted`
and `Feasible` sum cases to be compared against each other.

The test cases for `Fitness` have also have a significant overhaul to
ensure that the various cases are arbitrarily created and compared
against each other.

Fixes #287