Relax restriction (#292)
* Relax the restriction from NEL to Foldable

The `choices` and `sample` functions currently take in NonEmptyList,
and result in `RVar[Option[List]]` as the final result. Because the
number of items to sample is provided as a positive integer, the
number may be larger or smaller than the input structure (hence the
result is a `List`).

Some uses have appeared where the user has a `List` structure and not
a `NonEmptyList` - base don the internal operation of the function,
this will not be problematic so a nicer alternative is to allow all
`Foldable` structures. Both `NonEmptyList` and `List` fulfil this
constraint and are now usable with the function.

* No need for `Option[List[A]]`, just use `Option[List[A]]`

* Ensure logic remains unchanged

* Update RVar documentation

* Corrections and cleanup - reorder tests

* Simplify a lot of logic and remove some redundant actions