Lens fixes and testing Fit cases (#291)
* Simplify lenses to prevent invalid access

Several lenses inadvertantly allowed data to be changed by the user
in potentially dangerous ways - theis loop hole has been closed and
the offending lenses have been replaced with safe alternatives.

* Update Fitness tests and code to check against other cases

The objective code was only considering data that _had_ a fitness
value already assigned. Unfortunately this is a partial solution so
this commit adds the needed changes to allow `Infeasible`, `Adjusted`
and `Feasible` sum cases to be compared against each other.

The test cases for `Fitness` have also have a significant overhaul to
ensure that the various cases are arbitrarily created and compared
against each other.

Fixes #287

* Correct formatting issues

* Use Getter and not plain function with implicit conversion

* Point and Solution should not be visible externally

The API should manage the state of the Position internally and not
expose the data constructors. As a result, some optics are no longer
valid (their targets are now private) and have simply been removed.

* Code formating and test correction

* Simplification of fitness logic

The logic for "how to do a comparison" should be within the
`Comparison` values (such as `dominance`), and not in the optimisation
scheme target (Min or Max)

* General cleanup in file

* Correct tests

* Formatting changes

* Rebased onto master, updated to match new HEAD

* Add required type ascription

Signed-off-by: Gary Pamparà <gpampara@gmail.com>