Better output format derivation (#295)
* Better output format derivation

CSV derivation had a bug where all the headers were not output
to the resulting file. The fields were not flattened in the same
way that the data values were. The derivation code has been modified
slightly to make derivation slightly nicer and the constraint on
Measurement has been removed.

* Simplify IO composition