Fitness functions should determine fitness viability (#303)
This is a complete oversight: the provided function to the `Eval`
constructors should contain the logic to determine if a fitness value
is "good" or "bad". This necessitates that the fitness function return
the _type_ of fitness (which the quantification represents) instead of
a plain raw value. Doing so will allow for proper comparisons to
occur, based on the logic within the quantification function
itself. Currently, all results are seemingly "feasible" which, is not
the desired behaviour.