Add nushell to the development environment
The nushell (while unique in it's own right) allows for working with
dataframe data sources (such as .csv and .parquet) files in the shell
without needing to provide additional tools involving Python/Spark
software stacks.

There is a caveat here that the dataframe operation is relatively new
so it might not work in all cases. For example, the compression used
in our parquet sink is not currently supported but it is hoped to be
included in subsequent releases. The `pqrs` alternative is also
available (provided that it compiles).