refactor!: Remove cilib.Environment to make API more explicit
The `cilib.Environment` type was originally used to try to make the
API for execution simpler. This unfortunately failed  and resulted in
a more complex internal behaviour within the library. As a result
`cilib.Environment` has been removed in favor of letting the `Step`
and `StepS` types require the relevant `cilib.Comparison` and
`cilib.Eval` instances. The result is a more explicit usage API and
means that the errors are more direct without the need for an
indirection as provided by `cilib.Environment`.

Migration is simple enough and can be observed within the example
algorithms. The change requires the removal of the `cilib.Environment`
and then requires a `(cilib.Comparison, Eval[NonEmptyVector])` to be
passed to the `ZPure.provide` function. Previously, the `provide`
required the `cilib.Environment` type.