Drop spire in favour of simpler code to achieve the same goal
spire is not used for all we hoped and only a fraction of the code
available within the library is used. To this end (and that the
typelevel projects seem to be both dropping in quality and in general
usage) we simply manage the needed conversion and generic interfaces
ourselves.